Added settings file "arma-gamer-profile\Scripts\XGUI-ExtendedGUI\settings" Some fixes

+++ trunk/make-distrib.bat
@@ -12,32 +12,32 @@ set DistribDir_modDir=%DistribDir%\ArmA2\@\ext-editor
     echo -- START CREATE DISTRIBUTION PACK --
     echo ------------------------------------
 
-    rmdir /S /Q "%DistribDir%"
-    del /Q "%DistribDir%"
+    if exist "%DistribDir%" rmdir /S /Q "%DistribDir%"
+    if exist "%DistribDir%" call :Die "die: error!"
 
-    if exist "%DistribDir%" (
-        echo Error!
-        exit
-    )
+    if not exist "%DistribDir%\ArmA2" mkdir "%DistribDir%\ArmA2"
+    if not exist "%DistribDir_modDir%\addons" mkdir "%DistribDir_modDir%\addons"
 
-    mkdir "%DistribDir%\ArmA2"
-    mkdir "%DistribDir_modDir%\addons"
-
-    echo xcopy /E /Y "%TargetAddonDir%" "%DistribDir_modDir%\addons"
+    echo xcopy addons files:
     xcopy /E /Y "%TargetAddonDir%" "%DistribDir_modDir%\addons"
 
-    del /Q "%DistribDir_modDir%\addons\*.log"
-    rmdir /S /Q "%DistribDir_modDir%\addons\log"
+    if exist "%DistribDir_modDir%\addons\log" rmdir /S /Q "%DistribDir_modDir%\addons\log"
 
-    copy "mod.cpp" "%DistribDir_modDir%\mod.cpp"
+    if exist "mod.cpp" copy "mod.cpp" "%DistribDir_modDir%\mod.cpp"
+    if exist "readme*.txt" copy "readme*.txt" "%DistribDir_modDir%\*"
 
-    del "%DistribDir%.7z"
-    if exist "%DistribDir%.7z" (
-        echo Error!
-        exit
-    )
+    xcopy /E /Y "arma-gamer-profile" "%DistribDir%\arma-gamer-profile\\*"
+
+    if exist "%DistribDir%.7z" del "%DistribDir%.7z"
+    if exist "%DistribDir%.7z" call :Die "die: error!"
 
     "%Z7%" a -r0 -t7z -mx9 %Z7param% -scsDOS -- "%DistribDir%.7z" "%DistribDir%\*"
 
+    if not exist "%DistribDir%.7z" call :Die "die: error creating '%DistribDir%.7z'!"
+
 goto :eof
 
+:Die
+    echo %~1
+    exit
+goto :eof
